update: NiGuiとプレイヤー機能の改善
* NiGuiの変更:
  - `GameScene::Update()`メソッドで、`directLightParam_`の色、方向、強度を変更。
  - `Player::Initialize()`メソッドで、移動速度を1から35.0fに変更。
  - `Player::ImGui()`メソッドで、`MoveSpeed`のドラッグ操作を削除。
  - `Player::UpdateMovement()`メソッドで、移動速度の計算を`statusCurrent_.getSpeed()`を使用するように変更。
  - `Player.h`で、`moveSpeed_`の初期値を35.0fに設定。

* wnd_chain.cppの変更:
  - `GUI_Chain::Initialize()`メソッドで、アイテムのオフセットをリセットする処理を追加。
  - `GUI_Chain::OnNotify()`メソッドで、ウィンドウの表示状態に応じた通知処理を追加。
  - `GUI_Chain::Update()`メソッドで、ウィンドウが閉じたことを通知する処理を追加。
  - `GUI_Chain::ShowChain()`メソッドで、キャンセルボタン押下時に以前の状態に戻す処理を追加。

* Controller.cppの変更:
  - `Controller::Initialize()`メソッドで、ウィンドウサイズに基づく`statebar_`の位置設定を変更。
  - `Controller::Update()`メソッドで、ウィンドウサイズに基づく`statebar_`の位置再設定を追加。

* wnd_setting.cppの変更:
  - `Wnd_Setting::Initialize()`メソッドで、ウィンドウサイズをクラスメンバー変数に設定。
  - `Wnd_Setting::Update()`メソッドで、背景スプライトの位置をウィンドウサイズに基づいて更新。

* MyGame.cppの変更:
  - `MyGame::Update()`メソッドで、ウィンドウサイズの変更通知を明確に記述。

* TitleScene.cppの変更:
  - `TitleScene::Update()`メソッドで、ESCキー押下時にゲーム終了の通知を追加。

* imgui.iniの変更:
  - ウィンドウの位置やサイズを調整。

* StatusReinforcement.jsonの変更:
  - ステータス効果の値を変更し、特に移動速度、体力、攻撃力、防御力、クリティカル率の値を増加。

+++ GameProject/MyGame/MyGame.cpp
@@ -233,7 +233,13 @@ void MyGame::Update()
         ToggleFullScreen();
         NiGui::SetWindowInfo({ WinApp::clientWidth, WinApp::clientHeight }, {});
         NiGui::SetClientSize({ WinApp::clientWidth, WinApp::clientHeight });
-        GameEventNotifier::GetInstance()->Notify("OnWindowResized", Vector2(WinApp::clientWidth, WinApp::clientHeight));
+        GameEventNotifier::GetInstance()->Notify(
+            "OnWindowResized",
+            Vector2(
+                static_cast<float>(WinApp::clientWidth), 
+                static_cast<float>(WinApp::clientHeight)
+            )
+        );
     }
 
 #ifdef _DEBUG
